# Optimise `Version` representation