@dcoutts 2016-09-28
Should extend the tests in `Cabal/tests/UnitTests/Distribution/Version.hs`. Currently they cover `VersionRange` properties, but now that we have a non-trivial representation then I think we need proper coverage of `Version` itself.

So the things we need are: round trip tests, and Ord tests. These should have good coverage of versions with 0-6 components, and the boundary condition for the max size of each component.