@hvr 2016-09-28
@dcoutts  fwiw, here's some tests I used while developing:

```hs
prop_VersionId :: [Int] -> Bool
prop_VersionId lst0 = (versionNumbers . mkVersion) lst == lst
  where
    lst = map abs lst0

prop_VersionEq :: Base.Version -> Base.Version -> Bool
prop_VersionEq v1 v2 = (==) v1 v2 == ((==) `on` mkVersion') v1 v2

prop_VersionOrd :: Base.Version -> Base.Version -> Bool
prop_VersionOrd v1 v2 = compare v1 v2 == (compare `on` mkVersion') v1 v2
```