@dcoutts 2016-09-29
@hvr those tests look sensible. Add them to the Version.hs test module. It may be necessary to use a newtype wrapper to use a different Arbitrary instance for [Int]. It should make sure we get coverage of the 0xffff boundary, and not get too many uselessly long lists.