@hvr 2016-10-01
@ezyang I want to run a few criterion tests, look at runtime profiling reports, and maybe look at Core to see what kind of performance regression we may run into