@ezyang 2016-10-01
Sounds good. Obviously you don't have to, but it would be really awesome if this performance work could be recorded in some sort of durable way (ideally, running in the CI!) Allocations may be a decent, stable metric in this respect.