@hvr 2016-10-01
@ezyang @dcoutts fyi, I've updated the PR and removed the 'WIP' tag